refactor: search depending on isMobile
