+++ kickFlip-project/src/components/cardList/cardList.tsx
@@ -1,15 +1,17 @@
 import { ProductProjected, TransformParams } from '@/types/types';
 import Card from '../card/card';
 import './cardList.css';
+import SearchForm from '../searchForm/searchForm';
 
 interface CardListProps {
     products: ProductProjected[];
     setCategories: React.Dispatch<React.SetStateAction<TransformParams>>;
     setFilterIsActive: React.Dispatch<React.SetStateAction<boolean>>;
     categories: TransformParams;
+    isMobile: boolean;
 }
 
-function CardList({ products, setCategories, categories, setFilterIsActive }: CardListProps): JSX.Element {
+function CardList({ products, setCategories, categories, setFilterIsActive, isMobile }: CardListProps): JSX.Element {
     const handleSelect = (event: React.ChangeEvent<HTMLSelectElement>) => {
         setCategories((prevCategories) => {
             return { ...prevCategories, sort: event.target.value };
@@ -21,16 +23,15 @@ function CardList({ products, setCategories, categories, setFilterIsActive }: Ca
             <div className="card-list-info-container">
                 <h2 className="card-list-title">{`The best kicks (${products.length})`}</h2>
                 <div className="card-list-options-container">
+                    {isMobile && <SearchForm setCategories={setCategories} />}
                     <button
                         type="button"
                         onClick={() => {
                             setFilterIsActive((prevState) => !prevState);
                         }}
                         className="card-list-hide-filter"
-                    >
-                        <span className="filter-svg" />
-                        Show/Hide filter
-                    </button>
+                        aria-label="show/hide filters"
+                    />
                     <select id="sort-select" className="card-list-sorting" onChange={handleSelect}>
                         <option value="">Sort by</option>
                         <option value="price asc">Price (Low-High)</option>
