+++ kickFlip-project/src/components/filterComponent/filterComponent.css
@@ -54,32 +54,6 @@
     text-decoration: underline;
 }
 
-.filter-search-form {
-    display: flex;
-    align-items: center;
-    gap: 0.4rem;
-    position: relative;
-}
-
-.filter-search {
-    padding: 1rem 1.6rem;
-    border: 1px solid var(--border-gray);
-    color: var(--deep-blue);
-    width: 100%;
-    font-size: 1.3rem;
-}
-
-.filter-search-button {
-    inline-size: 2rem;
-    block-size: 2rem;
-    background-image: url('../../../public/search.svg');
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: contain;
-    position: absolute;
-    right: 1rem;
-}
-
 .filter-wrapper {
     display: flex;
     flex-direction: column;
