+++ kickFlip-project/src/pages/products/productsPage.css
@@ -31,7 +31,13 @@
     display: flex;
 }
 
-@media (width < 500px) {
+@media (width < 900px) {
+    .products {
+        display: flex;
+    }
+}
+
+@media (width < 600px) {
     .categories {
         grid-template-columns: 1fr;
         grid-template-rows: repeat(3, 1fr);
